% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrettyOutput.R
\name{printPrimaryCriteria}
\alias{printPrimaryCriteria}
\title{Print primary criteria}
\usage{
printPrimaryCriteria(obj, removeDescription = FALSE)
}
\arguments{
\item{obj}{Cohort object outputted from \code{CirceR} to be printed
in which case \code{json} is used.}

\item{removeDescription}{Currently not used (TODO fix)}
}
\description{
Outputs primary criteria into human-readable \code{markdown}
}
